<feature> fix: role 添加一个Box Collider/ isTriggle类型的用于是否在地面的检测

+++ Assets/Scripts_Runtime/Application_UI/HUD/HUD_HpBar.cs
@@ -15,13 +15,14 @@ namespace Act {
             transform.forward = forward;
         }
 
-        public void Update_Tick(int hp, Vector3 pos, float headOffset,Vector3 forward) {
+        public void Update_Tick(int hp, Vector3 pos, float headOffset, Vector3 forward) {
             // Hp
             if (hp <= 0) {
                 Destroy(gameObject);
                 return;
             }
-            hpBar.fillAmount = hp / hpMax;
+
+            hpBar.fillAmount = (float)hp / hpMax;
 
             // Pos 
             transform.position = pos + Vector3.up * headOffset;
